# openapi.json
{
  "components": {
    "schemas": {
      "ActionItem": {
        "properties": {
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "task": {
            "minLength": 1,
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "task"
        ],
        "title": "ActionItem",
        "type": "object"
      },
...
      "ApiErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "ArtifactResponse": {
        "properties": {
          "content_sha256": {
            "title": "Content Sha256",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "content_sha256",
          "schema_version",
          "revision",
          "created_at"
        ],
        "title": "ArtifactResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_revision": {
            "title": "Input Revision",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "settings_fingerprint": {
            "title": "Settings Fingerprint",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "attempts",
          "input_revision",
          "settings_fingerprint",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "MeetingSummary": {
        "properties": {
          "action_items": {
            "items": {
              "$ref": "#/components/schemas/ActionItem"
            },
            "title": "Action Items",
            "type": "array"
          },
          "decisions": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Decisions",
            "type": "array"
          },
          "discussion": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Discussion",
            "type": "array"
          },
          "open_questions": {
            "items": {
              "minLength": 1,
              "type": "string"
            },
            "title": "Open Questions",
            "type": "array"
          },
          "purpose": {
            "minLength": 1,
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "purpose",
          "discussion",
          "decisions",
          "action_items",
          "open_questions"
        ],
        "title": "MeetingSummary",
        "type": "object"
      },
...
      "RecordingDetailResponse": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactResponse"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Jobs",
            "type": "array"
          },
          "recording": {
            "$ref": "#/components/schemas/RecordingItem"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentResponse"
            },
            "title": "Segments",
            "type": "array"
          },
          "speakers": {
            "items": {
              "$ref": "#/components/schemas/RecordingSpeakerResponse"
            },
            "title": "Speakers",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MeetingSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "recording",
          "speakers",
          "segments",
          "artifacts",
          "jobs",
          "summary"
        ],
        "title": "RecordingDetailResponse",
        "type": "object"
      },
      "RecordingItem": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "category_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Confidence"
          },
          "category_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Reason"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "needs_speaker_review": {
            "title": "Needs Speaker Review",
            "type": "boolean"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/RecordingStatus"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "original_name",
          "duration_ms",
          "status",
          "category",
          "category_confidence",
          "category_reason",
          "needs_speaker_review",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "RecordingItem",
        "type": "object"
      },
      "RecordingListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecordingItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page_size": {
            "default": 50,
            "title": "Page Size",
            "type": "integer"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/RecordingStatus"
            },
            "title": "Status Counts",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "status_counts"
        ],
        "title": "RecordingListResponse",
        "type": "object"
      },
      "RecordingSpeakerResponse": {
        "properties": {
          "clip_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip Error Code"
          },
          "clip_status": {
            "enum": [
              "pending",
              "ready",
              "insufficient",
              "failed"
            ],
            "title": "Clip Status",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "local_speaker_id": {
            "title": "Local Speaker Id",
            "type": "string"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "representative_clip_artifact_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Representative Clip Artifact Id"
          },
          "representative_clip_end_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Representative Clip End Ms"
          },
          "representative_clip_start_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Representative Clip Start Ms"
          },
          "segment_count": {
            "title": "Segment Count",
            "type": "integer"
          },
          "speaker_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Name"
          },
          "speaker_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Score"
          },
          "speaker_source": {
            "enum": [
              "manual",
              "auto",
              "unresolved"
            ],
            "title": "Speaker Source",
            "type": "string"
          }
        },
        "required": [
          "local_speaker_id",
          "person_id",
          "speaker_name",
          "speaker_source",
          "speaker_score",
          "segment_count",
          "duration_ms",
          "clip_status",
          "clip_error_code",
          "representative_clip_artifact_id",
          "representative_clip_start_ms",
          "representative_clip_end_ms"
        ],
        "title": "RecordingSpeakerResponse",
        "type": "object"
      },
      "RecordingStatus": {
        "enum": [
          "DISCOVERED",
          "TRANSCRIBING",
          "SPEAKER_REVIEW",
          "CLASSIFYING",
          "READY_FOR_SUMMARY",
          "SUMMARIZING",
          "COMPLETED",
          "FAILED"
        ],
        "title": "RecordingStatus",
        "type": "string"
      },
...
      "SegmentResponse": {
        "properties": {
          "assignment_status": {
            "enum": [
              "assigned",
              "overlap",
              "unassigned"
            ],
            "title": "Assignment Status",
            "type": "string"
          },
          "end_ms": {
            "title": "End Ms",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "local_speaker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Speaker Id"
          },
          "overlapping_speaker_ids": {
            "items": {
              "type": "string"
            },
            "title": "Overlapping Speaker Ids",
            "type": "array"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Id"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "speaker_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Name"
          },
          "speaker_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker Score"
          },
          "speaker_source": {
            "enum": [
              "manual",
              "auto",
              "unresolved"
            ],
            "title": "Speaker Source",
            "type": "string"
          },
          "start_ms": {
            "title": "Start Ms",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "start_ms",
          "end_ms",
          "local_speaker_id",
          "assignment_status",
          "overlapping_speaker_ids",
          "person_id",
          "speaker_name",
          "speaker_source",
          "speaker_score",
          "text",
          "revision"
        ],
        "title": "SegmentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Voice to Document API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/recordings": {
      "get": {
        "operationId": "list_recordings_api_recordings_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RecordingStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Recordings",
        "tags": [
          "recordings"
        ]
      }
    },
    "/api/recordings/{recording_id}": {
      "get": {
        "operationId": "recording_detail_api_recordings__recording_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "recording_id",
            "required": true,
            "schema": {
              "title": "Recording Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recording Detail",
        "tags": [
          "recordings"
        ]
      }
    },
...
  }
}